requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.5
# sqlalchemy==2.0.9 //  Se instaló la version 1.4.48 por conflicto con graphene_sqlalchemy
flask_cors==3.0.10
flask_graphql==2.0.1
graphene==2.1.9
graphene_sqlalchemy==2.3.0